added readme and analytical files

+++ Algorithms/BubbleSort.cpp
@@ -3,8 +3,14 @@
 //
 
 #include <algorithm>
+#include <cstdint>
 #include "BubbleSort.h"
 
+/**
+ * Regular time-efficient bubble sort algorithm.
+ * @param array array to sort
+ * @param n array size
+ */
 void BubbleSort::bubbleSortTime(int *array, int n) {
     for (int i = 0; i < n - 1; ++i) {
         for (int j = 0; j < n - i - 1; ++j) {
@@ -15,6 +21,11 @@ void BubbleSort::bubbleSortTime(int *array, int n) {
     }
 }
 
+/**
+ * Time-efficient bubble sort algorithm with first Iverson rule.
+ * @param array array to sort
+ * @param n array size
+ */
 void BubbleSort::firstIversonBubbleSortTime(int *array, int n) {
     for (int i = 0; i < n - 1; ++i) {
         bool flag = true;
@@ -30,6 +41,11 @@ void BubbleSort::firstIversonBubbleSortTime(int *array, int n) {
     }
 }
 
+/**
+ * Time-efficient bubble sort algorithm with first and second Iverson rule.
+ * @param array array to sort
+ * @param n array size
+ */
 void BubbleSort::secondIversonBubbleSortTime(int *array, int n) {
     int prev = n - 1;
     for (int i = 0; i < n - 1; ++i) {
@@ -47,8 +63,14 @@ void BubbleSort::secondIversonBubbleSortTime(int *array, int n) {
     }
 }
 
-int BubbleSort::bubbleSortOperations(int *array, int n) {
-    int counter = 0;
+/**
+ * Regular operations counting bubble sort algorithm.
+ * @param array array to sort
+ * @param n array size
+ * @return number of performed elementary operations
+ */
+int64_t BubbleSort::bubbleSortOperations(int *array, int n) {
+    int64_t counter = 0;
     for (int i = 0; i < n - 1; ++i) {
         for (int j = 0; j < n - i - 1; ++j) {
             if (array[j] > array[j + 1]) {
@@ -63,8 +85,14 @@ int BubbleSort::bubbleSortOperations(int *array, int n) {
     return counter;
 }
 
-int BubbleSort::firstIversonBubbleSortOperations(int *array, int n) {
-    int counter = 0;
+/**
+ * Operations counting bubble sort algorithm with first Iverson rule.
+ * @param array array to sort
+ * @param n array size
+ * @return number of performed elementary operations
+ */
+int64_t BubbleSort::firstIversonBubbleSortOperations(int *array, int n) {
+    int64_t counter = 0;
     for (int i = 0; i < n - 1; ++i) {
         bool flag = true;
         for (int j = 0; j < n - i - 1; ++j) {
@@ -85,8 +113,14 @@ int BubbleSort::firstIversonBubbleSortOperations(int *array, int n) {
     return counter;
 }
 
-int BubbleSort::secondIversonBubbleSortOperations(int *array, int n) {
-    int counter = 0;
+/**
+ * Operations counting bubble sort algorithm with first and second Iverson rule.
+ * @param array array to sort
+ * @param n array size
+ * @return number of performed elementary operations
+ */
+int64_t BubbleSort::secondIversonBubbleSortOperations(int *array, int n) {
+    int64_t counter = 0;
     int prev = n - 1;
     for (int i = 0; i < n - 1; ++i) {
         int bound = prev;
