+++ Algorithms/HeapSort.h
@@ -4,15 +4,17 @@
 
 #pragma once
 
+#include <cstdint>
+
 struct HeapSort {
     static void heapSortTime(int *array, int n);
 
-    static int heapSortOperations(int *array, int n);
+    static int64_t heapSortOperations(int *array, int n);
 
 private:
     static void heapifyTime(int *array, int n, int i);
 
-    static void heapifyOperations(int *array, int n, int i, int &counter);
+    static void heapifyOperations(int *array, int n, int i, int64_t &counter);
 };
