+++ Algorithms/InsertionSort.h
@@ -9,14 +9,14 @@ struct InsertionSort {
 
     static void binaryInsertionSortTime(int *array, int n);
 
-    static int insertionSortOperations(int *array, int n);
+    static int64_t insertionSortOperations(int *array, int n);
 
-    static int binaryInsertionSortOperations(int *array, int n);
+    static int64_t binaryInsertionSortOperations(int *array, int n);
 
 private:
     static int binarySearchTime(int left, int right, int target, const int *array);
 
-    static int binarySearchOperations(int left, int right, int target, const int *array, int &counter);
+    static int binarySearchOperations(int left, int right, int target, const int *array, int64_t &counter);
 };
