+++ Algorithms/ShellSort.cpp
@@ -30,8 +30,8 @@ void ShellSort::ciuraSortTime(int *array, int n) {
     }
 }
 
-int ShellSort::shellSortOperations(int *array, int n) {
-    int counter = 0;
+int64_t ShellSort::shellSortOperations(int *array, int n) {
+    int64_t counter = 0;
     for (int gap = n / 2; gap > 0; gap /= 2) {
         for (int i = gap; i < n; ++i) {
             int value = array[i];
@@ -49,8 +49,8 @@ int ShellSort::shellSortOperations(int *array, int n) {
     return counter;
 }
 
-int ShellSort::ciuraSortOperations(int *array, int n) {
-    int counter = 0;
+int64_t ShellSort::ciuraSortOperations(int *array, int n) {
+    int64_t counter = 0;
     for (int gap : ciuraSequence) {
         for (int i = gap; i < n; ++i) {
             int value = array[i];
