+++ Algorithms/StableCountingSort.h
@@ -4,14 +4,16 @@
 
 #pragma once
 
+#include <cstdint>
+
 struct StableCountingSort {
     static void stableCountingSortTime(int *array, int n);
 
-    static int stableCountingSortOperations(int *array, int n);
+    static int64_t stableCountingSortOperations(int *array, int n);
 
 private:
     static int getMaxValueTime(int *array, int n);
 
-    static int getMaxValueOperations(int *array, int n, int &counter);
+    static int getMaxValueOperations(int *array, int n, int64_t &counter);
 };
